# openapi.yaml
openapi: 3.0.0
info:
  title: 'LookDesk API'
  version: 1.0.0
servers:
  -
    url: 'https://api.lookdesk.ai'
    description: 'Production API'
paths:
  /api/admin/plans:
    get:
      tags:
        - 'Admin Plans'
      operationId: 4b58fcd6d13184c5bf03c38a882f8a8e
      security:
        -
          sanctum: []
    post:
      tags:
        - 'Admin Plans'
      summary: 'Create plan with prices and features'
      operationId: 74e431da7a2c9aea218a525ed2bff36a
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
            example:
              code: starter
              name: Starter
              description: 'Perfect for small teams'
              status: available
              is_active: true
              prices:
                -
                  period: monthly
                  price: 30
                  per_user: false
                  trial_days: 30
                -
                  period: yearly
                  price: 300
                  per_user: false
                  trial_days: 30
              features:
                -
                  title: '30-days free trial'
                  sort: 1
                -
                  title: 'Up to 5 users'
                  sort: 2
                -
                  title: 'Live Chat'
                  sort: 3
      responses:
        '201':
          description: Created
        '422':
          description: 'Validation error'
      security:
        -
          sanctum: []
  '/api/admin/plans/{id}':
    get:
      tags:
        - 'Admin Plans'
      operationId: f89632d9fd37cb1716736560a98fa6b4
      security:
        -
          sanctum: []
    put:
      tags:
        - 'Admin Plans'
      summary: 'Update plan with prices and features'
      operationId: ea723497692bf628df359167405eec78
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
            example:
              name: Starter
              description: 'Perfect for small teams'
              status: available
              is_active: true
              prices:
                -
                  id: 1
                  period: monthly
                  price: 35
                  per_user: false
                  trial_days: 14
              features:
                -
                  id: 5
                  title: 'Live Chat'
                  sort: 1
      responses:
        '200':
          description: Updated
        '404':
          description: 'Not found'
      security:
        -
          sanctum: []
    delete:
      tags:
        - 'Admin Plans'
      operationId: 01e1d972c73e5b9491dc84936ed716e8
      security:
        -
          sanctum: []
  /api/auth/login:
    post:
      tags:
        - Auth
      summary: 'Login user'
      operationId: dd51a484b7fa0f97c7680fdbde14cf5f
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - email
                - password
              properties:
                email:
                  type: string
                password:
                  type: string
              type: object
      responses:
        '200':
          description: Success
        '401':
          description: 'Invalid credentials'
  /api/auth/register:
    post:
      tags:
        - Auth
      summary: 'Register user'
      operationId: c7bb21a0d87049e161b8e4c9b0d909f2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - name
                - email
                - password
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
              type: object
      responses:
        '201':
          description: 'User created'
  /api/auth/me:
    get:
      tags:
        - Auth
      summary: 'Get current user'
      operationId: b5de204f5e452ef51327c691aa99e77c
      responses:
        '200':
          description: 'Current user'
      security:
        -
          sanctum: []
  /api/auth/logout:
    post:
      tags:
        - Auth
      summary: Logout
      operationId: df9bdc6c7a6d9f20b983526b2ec0f698
      responses:
        '200':
          description: 'Logged out'
      security:
        -
          sanctum: []
...
components:
...
  securitySchemes:
    sanctum:
      type: http
      bearerFormat: JWT
      scheme: bearer
...